requirements.txt
# OpenADMET Project Requirements
# Machine learning and cheminformatics dependencies for ADMET prediction

# Core Data Science & Machine Learning
xgboost>=1.7.0
scikit-learn>=1.0.0
pandas>=1.5.0
numpy>=1.21.0

# Cheminformatics
rdkit>=2022.09.1

# Visualization
matplotlib>=3.5.0
seaborn>=0.12.0

# Deep Learning (for Chemprop)
torch>=2.0.0
lightning>=2.0.0

# Utilities
tqdm>=4.65.0
pyyaml>=6.0.0
jupyter>=1.0.0
ipykernel>=6.0.0
ipywidgets>=8.0.0

# Data Loading (for Hugging Face datasets)
huggingface-hub>=0.16.0
datasets>=2.14.0